ml/requirements.txt
pandas
pymysql
python-dotenv
scikit-learn
jieba
joblib
openpyxl
